Add serialization of `OathGame`s to save strings.
This now allows transferring a chronicle from the real world to TTS:
manually construct an `OathGame` object and then use `serializOathGame`
to get a string that can be loaded into TTS.

For every TTS save string, we can now parse the string into an
`OathGame` object and then serialize back to the original string.
`inverseParse.js` was added as a mini test suit to verify this
property against known strings.

Some small bugs in the parser were fixed in the process:

- Suit order was being parsed incorrectly.
- Previous game winner was always Brown and did not support Purple in enum.
- Previous game winner and exile status were looking at the wrong
offsets.